grow-478 created contentfulUtils function, exporting it and using it in HomepageStatistics.vue
